PropertyPanelからDrawAreaへの値渡しをいちいち変更するのがめんどくさくなってきたのでPropertyPanelを分割
